во всех классах изменил тип переменных с int на double. реализован счёт десятичных чисел.

+++ src/Calculation/Arithmetic/Addition.java
@@ -1,12 +1,12 @@
 package Calculation.Arithmetic;
 
 public class Addition extends Calculate {
-    public Addition(int a, int b, String operation) {
+    public Addition(double a, double b, String operation) {
         super(a, b, operation);
     }
 
-    public int resultOperation(int a, int b, String operation){
-        int resultOperatioin = a + b;
+    public double resultOperation(double a, double b, String operation){
+        double resultOperatioin = a + b;
         System.out.println(resultOperatioin);
         return resultOperatioin;
     }
