+++ src/Calculation/Arithmetic/Subtraction.java
@@ -1,12 +1,12 @@
 package Calculation.Arithmetic;
 
 public class Subtraction extends Calculate {
-    public Subtraction(int a, int b, String operation) {
+    public Subtraction(double a, double b, String operation) {
         super(a, b, operation);
     }
     @Override
-    public int resultOperation(int a, int b, String operation){
-        int resultOperatioin = a - b;
+    public double resultOperation(double a, double b, String operation){
+        double resultOperatioin = a - b;
         System.out.println(resultOperatioin);
         return resultOperatioin;
     }
